python_scripts/requirements.txt
pillow
requests
imageio
numpy
matplotlib
mycolorpy
geopandas
svg_path_transform
svg.path
svgpathtools
bcolors==1.0.2
pyvista
trimesh
beautifulsoup4
rich
numba
scikit-image
mnemonic
qahirah
randomcolor
perlin-noise
noise
opencv-python
pyfqmr
termcolor
distutils
svgutils